requirements.txt
Flask==2.3.2
flask-cors==3.0.10
pymongo[srv]==4.4.1
dnspython==2.4.2
python-dotenv==1.0.1
Flask-Session==0.8.0
bcrypt==4.0.1
flask-login==0.6.3
sentence-transformers==2.2.2
huggingface-hub==0.11.0 # Added this line
# faiss-cpu==1.12.0
# Pin numpy to a wheel-known version for Windows to avoid local build; adjust if you use conda
numpy==2.3.4
# Optional / recommended for vector search and scheduling
sentence-transformers==2.2.2
# If you want FAISS on CPU: uncomment the line below and install matching wheel for your platform
# faiss-cpu==1.12.0
apscheduler==3.10.1
requests==2.31.0
pytest==7.4.0
Flask-Limiter==2.11.0